pyproject.toml
[tool.poetry]
name = "file-uploader"
version = "0.1.0"
description = ""
authors = ["Егор Лоскутов <egorkaloskutov@yandex.ru>"]
readme = "README.md"
packages = [{include = "file_uploader"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "3.14.0"
djoser = "^2.2.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
drf-yasg = "^1.21.7"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

